Move datastore to upper package
