+++ server/tracedb/datastore/datastore_repository_test.go
@@ -6,10 +6,10 @@ import (
 	"testing"
 
 	"github.com/gorilla/mux"
+	datastore "github.com/kubeshop/tracetest/server/datastore"
 	"github.com/kubeshop/tracetest/server/pkg/id"
 	"github.com/kubeshop/tracetest/server/resourcemanager"
 	rmtests "github.com/kubeshop/tracetest/server/resourcemanager/testutil"
-	datastore "github.com/kubeshop/tracetest/server/tracedb/datastore"
 	"github.com/stretchr/testify/require"
 )
