+++ server/tracedb/jaegerdb.go
@@ -6,12 +6,12 @@ import (
 	"io"
 	"strings"
 
+	"github.com/kubeshop/tracetest/server/datastore"
 	pb "github.com/kubeshop/tracetest/server/internal/proto-gen-go/api_v3"
 	"github.com/kubeshop/tracetest/server/model"
 	"github.com/kubeshop/tracetest/server/pkg/id"
 	"github.com/kubeshop/tracetest/server/tracedb/connection"
 	"github.com/kubeshop/tracetest/server/tracedb/datasource"
-	"github.com/kubeshop/tracetest/server/tracedb/datastore"
 	"github.com/kubeshop/tracetest/server/traces"
 	v1 "go.opentelemetry.io/proto/otlp/trace/v1"
 	"google.golang.org/grpc"
